move GearBlock & message_queued_event_of to the common module
